+++ CHANGES.rst
@@ -50,6 +50,8 @@ datamodels
 
 - Update rscd model to increase the size of group_skip_table to allow FASTR1, SLOWR1, FASTR100 [#6776]
 
+- Correcting the default ZEROFRAME allocation. [#6791]
+
 extract_1d
 ----------